+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 HTTP
 
 * Make chunk_encode public
+* Add write, async_write, operator<< for message_headers
 
 WebSocket